added observation graph mechanism based on distance
