+++ experiments/drone_simulation.py
@@ -1,41 +1,29 @@
 from drone_algorithms import greedy_decision
+from utils import compute_adjacency_matrix, worker, individual_drone_score, plot_drone_movements
 import random
-import matplotlib.pyplot as plt
-import numpy as np
 import concurrent.futures
+import numpy as np
 
 class Drone:
-    def __init__(self, decision_function, position=(0.0,0.0)):
+    def __init__(self, decision_function, position=(0.0,0.0), observation_radius=50):  # default observation radius
         self.x, self.y = position
         self.intended_direction = (0, 0)
         self.decision_function = decision_function
         self.current_cost = float('inf')
+        self.observation_radius = observation_radius
+
+    def visible_drones(self, all_drones):
+        """ Return a list of drones that are visible to the current drone. """
+        return [d for d in all_drones if np.sqrt((self.x - d.x)**2 + (self.y - d.y)**2) <= self.observation_radius]
 
     def calculate_direction(self, all_drones, drone_index, D, **kwargs):
-        self.intended_direction = self.decision_function(self, all_drones, drone_index, D, **kwargs)
+        adjacency_matrix = compute_adjacency_matrix(all_drones)
+        visible_drones = [all_drones[j] for j in range(len(all_drones)) if adjacency_matrix[drone_index][j] == 1]
+        self.intended_direction = self.decision_function(self, visible_drones, drone_index, D, **kwargs)
 
     def move(self):
         self.x += self.intended_direction[0]
         self.y += self.intended_direction[1]
-
-def worker(drone, all_drones, idx, D, distance_to_origin_weight, epsilon):
-    drone.calculate_direction(all_drones, idx, D, distance_to_origin_weight=distance_to_origin_weight, epsilon=epsilon)
-    drone.move()  # move the drone after calculation
-    return drone
-
-def individual_drone_score(drone, all_drones, drone_index, D):
-    score = 0
-    
-    # Drone's distance from the origin
-    score += sum([np.sqrt(d.x**2 + d.y**2) for d in all_drones]) / len(all_drones)
-    
-    # Sum of squared differences from the desired distances
-    for j, other_drone in enumerate(all_drones):
-        if j != drone_index:
-            dist = np.sqrt((drone.x - other_drone.x)**2 + (drone.y - other_drone.y)**2)
-            score += (dist - D[drone_index][j])**2
-    
-    return score
 
 def simulate_environment(N, D, initial_positions=None, decision_function=greedy_decision, distance_to_origin_weight=1, epsilon=0.1, verbose=True):
     global DISTANCE_TO_ORIGIN_WEIGHT
@@ -69,21 +57,8 @@ def simulate_environment(N, D, initial_positions=None, decision_function=greedy_
         with concurrent.futures.ThreadPoolExecutor() as executor:
             drones = list(executor.map(worker, drones, [drones]*N, range(N), [D]*N, [distance_to_origin_weight]*N, [epsilon]*N))
 
-
     if verbose:
-        # Plotting
-        plt.figure(figsize=(5,5))
-        for iteration, positions in enumerate(positions_history):
-            for idx, (x, y) in enumerate(positions):
-                color = ['red', 'blue', 'green', 'orange', 'purple', 'brown'][idx % 6]
-                plt.plot(x, y, color, label="Drone" + str(idx) if iteration == 1 else "", alpha=0.1)
-                plt.text(x, y, str(iteration), color=color, fontsize=8, ha='center', va='center')
-
-        plt.title('Drone Movements Over Iterations')
-        plt.xlabel('X-coordinate')
-        plt.ylabel('Y-coordinate')
-        plt.grid(True)
-        plt.show()
+        plot_drone_movements(positions_history)
 
     # Calculate average score
     avg_score = sum([individual_drone_score(drone, drones, idx, D) for idx, drone in enumerate(drones)]) / N
